CHORE: restructure Arcus.Security.Core
Restructure the security core project by using only these categories:
`Caching`, `Configuration`, and all other files at the root of the
project.

Relates to #100
