+++ src/Arcus.Security.All/Arcus.Security.All.csproj
@@ -26,7 +26,6 @@
     <ProjectReference Include="..\Arcus.Security.Core\Arcus.Security.Core.csproj" />
     <ProjectReference Include="..\Arcus.Security.Providers.AzureKeyVault\Arcus.Security.Providers.AzureKeyVault.csproj" />
     <ProjectReference Include="..\Arcus.Security.Secrets.AzureKeyVault\Arcus.Security.Secrets.AzureKeyVault.csproj" />
-    <ProjectReference Include="..\Arcus.Security.Secrets.Core\Arcus.Security.Secrets.Core.csproj" />
   </ItemGroup>
 
 </Project>
