+++ src/Arcus.Security.Secrets.AzureKeyVault/Arcus.Security.Secrets.AzureKeyVault.csproj
@@ -22,7 +22,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Arcus.Security.Providers.AzureKeyVault\Arcus.Security.Providers.AzureKeyVault.csproj" />
-    <ProjectReference Include="..\Arcus.Security.Secrets.Core\Arcus.Security.Secrets.Core.csproj" />
   </ItemGroup>
 
 </Project>
