+++ src/Arcus.Security.Tests.Unit/Core/CachedSecretProviderTests.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Threading.Tasks;
+using Arcus.Security.Core;
+using Arcus.Security.Core.Caching;
 using Arcus.Security.Core.Caching.Configuration;
-using Arcus.Security.Secrets.Core.Caching;
-using Arcus.Security.Secrets.Core.Interfaces;
-using Arcus.Security.Secrets.Core.Models;
 using Arcus.Security.Tests.Unit.Core.Stubs;
 using Microsoft.Extensions.Caching.Memory;
 using Xunit;
