+++ src/Arcus.Security.Tests.Unit/Core/Stubs/TestSecretProviderStub.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Arcus.Security.Secrets.Core.Exceptions;
-using Arcus.Security.Secrets.Core.Interfaces;
-using Arcus.Security.Secrets.Core.Models;
+using Arcus.Security.Core;
 
 namespace Arcus.Security.Tests.Unit.Core.Stubs
 {
