+++ src/Arcus.Security.Tests.Unit/KeyVault/KeyVaultSecretProviderTests.cs
@@ -1,6 +1,6 @@
 ﻿using Arcus.Security.Providers.AzureKeyVault.Configuration;
 using Arcus.Security.Secrets.AzureKeyVault;
-using Arcus.Security.Secrets.Core.Models;
+using Arcus.Security.Core;
 using Arcus.Security.Tests.Unit.KeyVault.Doubles;
 using Microsoft.Azure.KeyVault.Models;
 using Microsoft.Rest;
