+++ src/Arcus.Security.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28922.388
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Secrets.Core", "Arcus.Security.Secrets.Core\Arcus.Security.Secrets.Core.csproj", "{A179BD58-7555-4F74-8EAF-CBD4F7BD689E}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Tests.Unit", "Arcus.Security.Tests.Unit\Arcus.Security.Tests.Unit.csproj", "{DACBC3C6-2894-4961-B327-746BC29229CC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Tests.Integration", "Arcus.Security.Tests.Integration\Arcus.Security.Tests.Integration.csproj", "{D185D73E-D617-4846-9CAB-12884C1266C5}"
@@ -19,8 +17,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Providers", "Providers", "{
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Secrets.AzureKeyVault", "Arcus.Security.Secrets.AzureKeyVault\Arcus.Security.Secrets.AzureKeyVault.csproj", "{473728CE-777E-4FC3-8164-6B603088C877}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{7E59A380-48DB-443B-9B47-05A4A175B8F3}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.Core", "Arcus.Security.Core\Arcus.Security.Core.csproj", "{EC4CBB2C-9C38-4D5C-B4C4-602F0756A61F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.Security.All", "Arcus.Security.All\Arcus.Security.All.csproj", "{ADB320AA-15FB-4FAB-A909-A1EE120BE7F9}"
@@ -31,10 +27,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A179BD58-7555-4F74-8EAF-CBD4F7BD689E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A179BD58-7555-4F74-8EAF-CBD4F7BD689E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A179BD58-7555-4F74-8EAF-CBD4F7BD689E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A179BD58-7555-4F74-8EAF-CBD4F7BD689E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DACBC3C6-2894-4961-B327-746BC29229CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DACBC3C6-2894-4961-B327-746BC29229CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DACBC3C6-2894-4961-B327-746BC29229CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -64,12 +56,10 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{A179BD58-7555-4F74-8EAF-CBD4F7BD689E} = {26FD7AFD-F12E-4F1F-965E-379065340D47}
 		{DACBC3C6-2894-4961-B327-746BC29229CC} = {C42FDAC3-5A31-4F39-BD92-C57C2CDAB2D2}
 		{D185D73E-D617-4846-9CAB-12884C1266C5} = {C42FDAC3-5A31-4F39-BD92-C57C2CDAB2D2}
 		{DD03DFBB-7D61-4FC9-B2D0-1BB800EA1ED9} = {DD66B1E8-3676-4C13-8C37-AEA80E1C21B7}
 		{473728CE-777E-4FC3-8164-6B603088C877} = {26FD7AFD-F12E-4F1F-965E-379065340D47}
-		{EC4CBB2C-9C38-4D5C-B4C4-602F0756A61F} = {7E59A380-48DB-443B-9B47-05A4A175B8F3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D3310D31-C37D-47F4-A6D3-325EE3806698}
